Addressing reviewer comments
