+++ test/unit/specs/components/Personalization/actions/move.spec.js
@@ -11,7 +11,7 @@ describe("Personalization::actions::move", () => {
     cleanUpDomChanges("move");
   });
 
-  it("should set personalized content", done => {
+  it("should move personalized content", () => {
     const collect = jasmine.createSpy();
     const modules = initRuleComponentModules(collect);
     const { move } = modules;
@@ -27,16 +27,10 @@ describe("Personalization::actions::move", () => {
     };
     const event = { elements, prehidingSelector: "#move" };
 
-    move(settings, event)
-      .then(() => {
-        expect(elements[0].style.left).toEqual("100px");
-        expect(elements[0].style.top).toEqual("100px");
-        expect(collect).toHaveBeenCalledWith(meta);
-        done();
-      })
-      .catch(() => {
-        fail("Should not fail");
-        done();
-      });
+    move(settings, event).then(() => {
+      expect(elements[0].style.left).toEqual("100px");
+      expect(elements[0].style.top).toEqual("100px");
+      expect(collect).toHaveBeenCalledWith(meta);
+    });
   });
 });
