+++ test/unit/specs/components/Personalization/actions/prependHtml.spec.js
@@ -15,7 +15,7 @@ describe("Personalization::actions::prependHtml", () => {
     cleanUpDomChanges("prependHtml");
   });
 
-  it("should prepend personalized content", done => {
+  it("should prepend personalized content", () => {
     const collect = jasmine.createSpy();
     const modules = initRuleComponentModules(collect);
     const { prependHtml } = modules;
@@ -36,20 +36,14 @@ describe("Personalization::actions::prependHtml", () => {
     };
     const event = { elements, prehidingSelector: "#prependHtml" };
 
-    prependHtml(settings, event)
-      .then(() => {
-        const result = selectNodes("ul#prependHtml li");
+    return prependHtml(settings, event).then(() => {
+      const result = selectNodes("ul#prependHtml li");
 
-        expect(result.length).toEqual(3);
-        expect(result[0].innerHTML).toEqual("1");
-        expect(result[1].innerHTML).toEqual("2");
-        expect(result[2].innerHTML).toEqual("3");
-        expect(collect).toHaveBeenCalledWith(meta);
-        done();
-      })
-      .catch(() => {
-        fail("Should not fail");
-        done();
-      });
+      expect(result.length).toEqual(3);
+      expect(result[0].innerHTML).toEqual("1");
+      expect(result[1].innerHTML).toEqual("2");
+      expect(result[2].innerHTML).toEqual("3");
+      expect(collect).toHaveBeenCalledWith(meta);
+    });
   });
 });
