+++ test/unit/specs/components/Personalization/actions/rearrange.spec.js
@@ -15,7 +15,7 @@ describe("Presonalization::actions::rearrange", () => {
     cleanUpDomChanges("rearrange");
   });
 
-  it("should rearrange elements when from < to", done => {
+  it("should rearrange elements when from < to", () => {
     const collect = jasmine.createSpy();
     const modules = initRuleComponentModules(collect);
     const { rearrange } = modules;
@@ -37,23 +37,17 @@ describe("Presonalization::actions::rearrange", () => {
     const settings = { content: { from: 0, to: 2 }, meta };
     const event = { elements, prehidingSelector: "#rearrange" };
 
-    rearrange(settings, event)
-      .then(() => {
-        const result = selectNodes("li");
+    return rearrange(settings, event).then(() => {
+      const result = selectNodes("li");
 
-        expect(result[0].textContent).toEqual("2");
-        expect(result[1].textContent).toEqual("3");
-        expect(result[2].textContent).toEqual("1");
-        expect(collect).toHaveBeenCalledWith(meta);
-        done();
-      })
-      .catch(() => {
-        fail("Should not fail");
-        done();
-      });
+      expect(result[0].textContent).toEqual("2");
+      expect(result[1].textContent).toEqual("3");
+      expect(result[2].textContent).toEqual("1");
+      expect(collect).toHaveBeenCalledWith(meta);
+    });
   });
 
-  it("should rearrange elements when from > to", done => {
+  it("should rearrange elements when from > to", () => {
     const collect = jasmine.createSpy();
     const modules = initRuleComponentModules(collect);
     const { rearrange } = modules;
@@ -75,19 +69,13 @@ describe("Presonalization::actions::rearrange", () => {
     const settings = { content: { from: 2, to: 0 }, meta };
     const event = { elements, prehidingSelector: "#rearrange" };
 
-    rearrange(settings, event)
-      .then(() => {
-        const result = selectNodes("li");
+    return rearrange(settings, event).then(() => {
+      const result = selectNodes("li");
 
-        expect(result[0].textContent).toEqual("3");
-        expect(result[1].textContent).toEqual("1");
-        expect(result[2].textContent).toEqual("2");
-        expect(collect).toHaveBeenCalledWith(meta);
-        done();
-      })
-      .catch(() => {
-        fail("Should not fail");
-        done();
-      });
+      expect(result[0].textContent).toEqual("3");
+      expect(result[1].textContent).toEqual("1");
+      expect(result[2].textContent).toEqual("2");
+      expect(collect).toHaveBeenCalledWith(meta);
+    });
   });
 });
