+++ test/unit/specs/components/Personalization/actions/resize.spec.js
@@ -11,7 +11,7 @@ describe("Personalization::actions::resize", () => {
     cleanUpDomChanges("resize");
   });
 
-  it("should set personalized content", done => {
+  it("should resize personalized content", () => {
     const collect = jasmine.createSpy();
     const modules = initRuleComponentModules(collect);
     const { resize } = modules;
@@ -27,16 +27,10 @@ describe("Personalization::actions::resize", () => {
     };
     const event = { elements, prehidingSelector: "#resize" };
 
-    resize(settings, event)
-      .then(() => {
-        expect(elements[0].style.width).toEqual("100px");
-        expect(elements[0].style.height).toEqual("100px");
-        expect(collect).toHaveBeenCalledWith(meta);
-        done();
-      })
-      .catch(() => {
-        fail("Should not fail");
-        done();
-      });
+    return resize(settings, event).then(() => {
+      expect(elements[0].style.width).toEqual("100px");
+      expect(elements[0].style.height).toEqual("100px");
+      expect(collect).toHaveBeenCalledWith(meta);
+    });
   });
 });
