+++ test/unit/specs/components/Personalization/actions/setImageSource.spec.js
@@ -11,7 +11,7 @@ describe("Personalization::actions::setImageSource", () => {
     cleanUpDomChanges("setImageSource");
   });
 
-  it("should swap image", done => {
+  it("should swap image", () => {
     const url = "http://foo.com/a.png";
     const collect = jasmine.createSpy();
     const modules = initRuleComponentModules(collect);
@@ -25,15 +25,9 @@ describe("Personalization::actions::setImageSource", () => {
     const settings = { content: "http://foo.com/b.png", meta };
     const event = { elements, prehidingSelector: "#setImageSource" };
 
-    setImageSource(settings, event)
-      .then(() => {
-        expect(elements[0].getAttribute("src")).toEqual("http://foo.com/b.png");
-        expect(collect).toHaveBeenCalledWith(meta);
-        done();
-      })
-      .catch(() => {
-        fail("Should not fail");
-        done();
-      });
+    return setImageSource(settings, event).then(() => {
+      expect(elements[0].getAttribute("src")).toEqual("http://foo.com/b.png");
+      expect(collect).toHaveBeenCalledWith(meta);
+    });
   });
 });
